Tidied up the solution and removed redundant console app

+++ VtlSoftware.StructuredLoggingConsoleApp.Net6/ICalculator.cs
@@ -3,9 +3,10 @@
     public interface ICalculator
     {
         #region Public Methods
+
         double Add(double a, double b);
 
-        double Divide(double a, double b);
+        int Divide(int a, int b);
 
         double Multiply(double a, double b);
